cleanup

+++ packages/chain/contracts/BlsVerify.sol
@@ -5,25 +5,25 @@ import { modexp_3064_fd54, modexp_c191_3f52 } from "./modexp.sol";
 
 contract BlsVerify {
     // Field order
-    uint256 constant N = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
+    uint constant N = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
 
     // Negated genarator of G2
-    uint256 constant nG2x1 = 11559732032986387107991004021392285783925812861821192530917403151452391805634;
-    uint256 constant nG2x0 = 10857046999023057135944570762232829481370756359578518086990519993285655852781;
-    uint256 constant nG2y1 = 17805874995975841540914202342111839520379459829704422454583296818431106115052;
-    uint256 constant nG2y0 = 13392588948715843804641432497768002650278120570034223513918757245338268106653;
+    uint constant nG2x1 = 11559732032986387107991004021392285783925812861821192530917403151452391805634;
+    uint constant nG2x0 = 10857046999023057135944570762232829481370756359578518086990519993285655852781;
+    uint constant nG2y1 = 17805874995975841540914202342111839520379459829704422454583296818431106115052;
+    uint constant nG2y0 = 13392588948715843804641432497768002650278120570034223513918757245338268106653;
 
     // sqrt(-3)
-    uint256 constant z0 = 0x0000000000000000b3c4d79d41a91759a9e4c7e359b6b89eaec68e62effffffd;
+    uint constant z0 = 0x0000000000000000b3c4d79d41a91759a9e4c7e359b6b89eaec68e62effffffd;
     // (sqrt(-3) - 1)  / 2
-    uint256 constant z1 = 0x000000000000000059e26bcea0d48bacd4f263f1acdb5c4f5763473177fffffe;
+    uint constant z1 = 0x000000000000000059e26bcea0d48bacd4f263f1acdb5c4f5763473177fffffe;
 
-    uint256 constant FIELD_MASK = 0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff;
-    uint256 constant SIGN_MASK = 0x8000000000000000000000000000000000000000000000000000000000000000;
-    uint256 constant ODD_NUM = 0x8000000000000000000000000000000000000000000000000000000000000000;
+    uint constant FIELD_MASK = 0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff;
+    uint constant SIGN_MASK = 0x8000000000000000000000000000000000000000000000000000000000000000;
+    uint constant ODD_NUM = 0x8000000000000000000000000000000000000000000000000000000000000000;
 
-    uint256 constant T24 = 0x1000000000000000000000000000000000000000000000000;
-    uint256 constant MASK24 = 0xffffffffffffffffffffffffffffffffffffffffffffffff;
+    uint constant T24 = 0x1000000000000000000000000000000000000000000000000;
+    uint constant MASK24 = 0xffffffffffffffffffffffffffffffffffffffffffffffff;
 
     bytes constant cipher_suite_domain = bytes('BNS_SIG_BNS256_XMD:SHA-256_SSWU');
 
@@ -35,9 +35,9 @@ contract BlsVerify {
         return true;
     }
 
-    function verifySignature(uint256[2] memory signature, uint256[4] memory pubkey, uint256[2] memory message) public view returns (bool) {
-        uint256[12] memory input = [signature[0], signature[1], nG2x1, nG2x0, nG2y1, nG2y0, message[0], message[1], pubkey[1], pubkey[0], pubkey[3], pubkey[2]];
-        uint256[1] memory out;
+    function verifySignature(uint[2] memory signature, uint[4] memory pubkey, uint[2] memory message) public view returns (bool) {
+        uint[12] memory input = [signature[0], signature[1], nG2x1, nG2x0, nG2y1, nG2y0, message[0], message[1], pubkey[1], pubkey[0], pubkey[3], pubkey[2]];
+        uint[1] memory out;
         bool success;
         // solium-disable-next-line security/no-inline-assembly
         assembly {
@@ -51,11 +51,11 @@ contract BlsVerify {
         return out[0] != 0;
     }
 
-    function hashToPoint(bytes memory message) public view returns (uint256[2] memory) {
-        uint256[2] memory u = hashToField(cipher_suite_domain, message);
-        uint256[2] memory p0 = mapToPointFT(u[0]);
-        uint256[2] memory p1 = mapToPointFT(u[1]);
-        uint256[4] memory bnAddInput;
+    function hashToPoint(bytes memory message) public view returns (uint[2] memory) {
+        uint[2] memory u = hashToField(cipher_suite_domain, message);
+        uint[2] memory p0 = mapToPointFT(u[0]);
+        uint[2] memory p1 = mapToPointFT(u[1]);
+        uint[4] memory bnAddInput;
         bnAddInput[0] = p0[0];
         bnAddInput[1] = p0[1];
         bnAddInput[2] = p1[0];
@@ -87,12 +87,12 @@ contract BlsVerify {
 
         return string(hexString);
     }
-    function hashToField(bytes memory domain, bytes memory messages) internal pure returns (uint256[2] memory) {
+    function hashToField(bytes memory domain, bytes memory messages) internal pure returns (uint[2] memory) {
         bytes memory _msg = expandMsgTo96(domain, messages);
-        uint256 z0;
-        uint256 z1;
-        uint256 a0;
-        uint256 a1;
+        uint z0;
+        uint z1;
+        uint a0;
+        uint a1;
         // solium-disable-next-line security/no-inline-assembly
         assembly {
             let p := add(_msg, 24)
@@ -109,14 +109,14 @@ contract BlsVerify {
         return [a0, a1];
     }
 
-    function mapToPointFT(uint256 _x) internal view returns (uint256[2] memory p) {
+    function mapToPointFT(uint _x) internal view returns (uint[2] memory p) {
         require(_x < N, "mapToPointFT: invalid field element");
-        uint256 x = _x;
+        uint x = _x;
         bool decision = isNonResidueFP(x);
-        uint256 a0 = mulmod(x, x, N);
+        uint a0 = mulmod(x, x, N);
         a0 = addmod(a0, 4, N);
-        uint256 a1 = mulmod(x, z0, N);
-        uint256 a2 = mulmod(a1, a0, N);
+        uint a1 = mulmod(x, z0, N);
+        uint a2 = mulmod(a1, a0, N);
         a2 = inverseFaster(a2);
         a1 = mulmod(a1, a1, N);
         a1 = mulmod(a1, a2, N);
@@ -170,10 +170,10 @@ contract BlsVerify {
     }
 
     function expandMsgTo96(bytes memory domain, bytes memory message) internal pure returns (bytes memory) {
-        uint256 t1 = domain.length;
+        uint t1 = domain.length;
         require(t1 < 256, "BLS: invalid domain length");
         // zero<64>|msg<var>|lib_str<2>|I2OSP(0, 1)<1>|dst<var>|dst_len<1>
-        uint256 t0 = message.length;
+        uint t0 = message.length;
         bytes memory msg0 = new bytes(t1 + t0 + 64 + 4);
         bytes memory out = new bytes(96);
         // b0
@@ -267,7 +267,7 @@ contract BlsVerify {
 
         return out;
     }
-    function isNonResidueFP(uint256 e) internal view returns (bool isNonResidue) {
+    function isNonResidueFP(uint e) internal view returns (bool isNonResidue) {
         bool callSuccess;
         // solium-disable-next-line security/no-inline-assembly
         assembly {
@@ -286,10 +286,10 @@ contract BlsVerify {
         require(callSuccess, "BLS: isNonResidueFP modexp call failed");
         return !isNonResidue;
     }
-    function inverseFaster(uint256 a) internal pure returns (uint256) {
+    function inverseFaster(uint a) internal pure returns (uint) {
         return modexp_3064_fd54.run(a);
     }
-    function sqrtFaster(uint256 xx) internal pure returns (uint256 x, bool hasRoot) {
+    function sqrtFaster(uint xx) internal pure returns (uint x, bool hasRoot) {
         x = modexp_c191_3f52.run(xx);
         hasRoot = mulmod(x, x, N) == xx;
     }
