+++ packages/chain/contracts/tokens/RegularERC20.sol
@@ -13,11 +13,11 @@ contract RegularERC20 is ERC20 {
     constructor(string memory name_, string memory symbol_) ERC20(name_, symbol_) {
         owner = msg.sender;
     }
-    function mint(address account, uint256 amount) public isOwner {
+    function mint(address account, uint amount) public isOwner {
         _mint(account, amount);
     }
 
-    function burn(address account, uint256 amount) public isOwner {
+    function burn(address account, uint amount) public isOwner {
         _burn(account, amount);
     }
 }
